Significant changes, removed cart class.
Each product added will be stored as a product class which is a struct.
Methods are added which checks for the received product type and calls
respective method which appends the product to the list. A custom signal
is then emitted which allows us to update the cart widget with the last
value.